adding v2.15.2

+++ src/library/stats/src/d2x2xk.c
@@ -1,3 +1,22 @@
+/*
+ *  R : A Computer Language for Statistical Data Analysis
+ *  Copyright (C) 2000-1   The R Core Team.
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program; if not, a copy is available at
+ *  http://www.r-project.org/Licenses/
+ */
+
 #include <R.h>
 #include <Rmath.h>
 
@@ -14,8 +33,8 @@ d2x2xk(Sint *k, double *m, double *n, double *t, double *d)
     c[0] = (double *) R_alloc(1, sizeof(double));
     c[0][0] = 1;
     for(i = 0; i < *k; i++) {
-	y = imax2(0, *t - *n);
-	z = imin2(*m, *t);
+	y = imax2(0,  (int)(*t - *n));
+	z = imin2((int)*m, (int)*t);
 	c[i + 1] = (double *) R_alloc(l + z - y + 1, sizeof(double));
 	for(j = 0; j <= l + z - y; j++)
 	    c[i + 1][j] = 0;
